# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Project Backyard API
  version: 1.0.0
  description: |
    人間とAIエージェントが共同で使うプロジェクト管理ツールの REST API。

    実装済みの範囲のみを記述する。設計の正本は `docs/ApiDesign.md`。
servers:
  # /healthcheck は唯一 /api/v1 の外にあるため（ApiDesign.md 2.11）、
  # ベースをルートに置き、paths 側に完全なパスを書く。
  - url: /
    description: 同一オリジン
...
paths:
...
  /api/v1/me/tokens:
    get:
      tags: [me]
      summary: 自分のアクセストークン一覧
      operationId: listMyTokens
      description: |
        CLI・スクリプト用の Bearer トークンを一覧する（ApiDesign.md 4.4.1）。
        必要権限は「本人」。

        **`token_type='api'` の行だけを返す。** ブラウザのセッションは現れない
        （本人が自分のセッションを見る画面を持たないため。GuiDesign.md 5.8）。

        **平文（`token`）は返さない。** 返すのは `token_prefix`（先頭8文字）だけで、
        再表示する経路は存在しない。

        **失効済みは返さない。期限切れは返す**（`status` が `expired`）——更新が要る
        ことに本人が気づく必要があり、上限5本を占めているためである。

        ページネーションも ETag も持たない（1人5本が上限）。
      responses:
        '200':
          description: 失効していないアクセストークン。`issued_at` の降順。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [me]
      summary: アクセストークンを発行
      operationId: createMyToken
      description: |
        自分のアクセストークンを発行する（ApiDesign.md 4.4.2）。必要権限は「本人」。

        **`token`（平文）はこの応答でのみ返る。** DB には SHA-256 のハッシュしか
        残らず、再表示する API は無い（DbDesign.md 6.2）。

        **`scopes` の語彙は権限カタログのキー**（`ticket.view` 等30件。Design.md
        6.4.2）。カタログに無い値は 422。空配列は「絞り込みなし」＝本人の実効権限
        そのままで、Phase 1 の画面は常にこれで発行する。

        **失効していないトークンが既に5本あると 409。** 期限切れも数える
        （`GET /me/tokens` が返す行と一致させるため）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTokenRequest'
      responses:
        '201':
          description: 発行したトークン。**`token` を返す唯一の応答**。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedAccessToken'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/CSRFFailed'
        '409':
          description: |
            失効していないトークンが既に5本ある（`conflict`）。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/me/agents:
    get:
      tags: [me]
      summary: 自分のエージェント一覧
      operationId: listMyAgents
      description: |
        自分の端末で動くクライアントを一覧する（ApiDesign.md 4.5.1）。
        必要権限は「本人」。

        **1件が表すのは「ある参加者の手元で動くクライアント1つ」である。** 人ではない。
        キーは（所有者・クライアント種別・プロジェクト）の3つ組で、同じ人が
        Claude Code と VS Code を使えば2件になる（Requirements.md 10.10.3）。

        **`token` は有効な1本だけを載せる。** 失効済みは返さず、期限切れは返す
        （`status` が `expired`）。有効なトークンが無ければ `null`。

        ページネーションも ETag も持たない（4.4.1 と同じ）。
      responses:
        '200':
          description: 自分のエージェント。`created_at` の降順。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyAgentList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [me]
      summary: エージェントを登録
      operationId: createMyAgent
      description: |
        自分のエージェントを登録する（ApiDesign.md 4.5.2）。必要権限は「本人」。

        **トークンは同時に発行しない**（応答の `token` は常に `null`）。登録と発行を
        分けるのは、再発行のときに同じ経路（`POST /me/agents/{id}/tokens`）を通す
        ためである。

        **`project_key` は自分がメンバーであるプロジェクトに限る。** それ以外は 422
        （`details[].code` は `not_found`）。エージェントの権限は所有者から導かれる
        ので、自分が入っていないプロジェクトのエージェントを作っても権限0件になる。

        **`trust_level` と `capabilities` は受け取らない**（Design.md 6.5）。
        昇格の材料が Phase 3 のため、既定値で作る。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: 登録したエージェント。`token` は `null`。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyAgent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/CSRFFailed'
        '409':
          description: |
            同じ（プロジェクト・クライアント種別・表示名）の組が既にある
            （`already_exists`）。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/me/agents/{id}:
    parameters:
      - $ref: '#/components/parameters/AgentID'
    patch:
      tags: [me]
      summary: エージェントを更新・無効化
      operationId: updateMyAgent
      description: |
        自分のエージェントを更新する（ApiDesign.md 4.5.4）。必要権限は「本人」。
        送られた項目だけを更新する。

        **`project_key` と `client_kind` は変えられない**（本要求に含まれない）。
        その2つが「どのクライアントがどのプロジェクトにつないでいるか」という
        1件の同一性そのものであり、付け替えると過去の操作の文脈が後から変わる。

        **`is_active: false` が無効化である。** 行は消さない（`comment.author_id`
        などが参照するため）。**無効化すると、そのエージェントのトークンも失効する。**

        **他人のエージェント・存在しない `id` は 404**（403 にしない。Design.md 6.4.5）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: 更新後のエージェント。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyAgent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/CSRFFailed'
        '404':
          description: |
            自分のエージェントとして見つからない（`not_found`）。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/me/agents/{id}/tokens:
    parameters:
      - $ref: '#/components/parameters/AgentID'
    post:
      tags: [me]
      summary: エージェント用トークンを発行
      operationId: createMyAgentToken
      description: |
        エージェント用の Bearer トークンを発行する（ApiDesign.md 4.5.3）。
        必要権限は「本人」。

        **`token`（平文）はこの応答でのみ返る。** DB には SHA-256 のハッシュしか
        残らず、再表示する API は無い（DbDesign.md 6.2）。

        **有効なトークンは1件につき1本。** 再発行すると、それまでの有効なトークンを
        失効させたうえで発行する（応答は 201。失効は暗黙に行い、監査には残る）。

        **`scopes` は選ばせない。** Design.md 6.5 の既定を常に入れる——
        `ticket.close` と `doc.edit` を既定から外してあるのが要点で、選ばせると
        その禁止が画面の作りに依存してしまう。**トークンは所有者の権限との積に
        なる**ため、既定をそのまま載せても所有者が持たない権限は付かない。

        発行するトークンは `project_id` を持ち、他プロジェクトへのアクセスは 404 になる
        （Design.md 6.4.5）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentTokenRequest'
      responses:
        '201':
          description: 発行したトークン。**`token` を返す唯一の応答**。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedAgentToken'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/CSRFFailed'
        '404':
          description: |
            自分のエージェントとして見つからない（`not_found`）。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            無効化されたエージェントには発行できない（`conflict`）。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/me/agents/{id}/tokens/{token_id}:
    parameters:
      - $ref: '#/components/parameters/AgentID'
      - $ref: '#/components/parameters/AgentTokenID'
    delete:
      tags: [me]
      summary: エージェント用トークンを失効
      operationId: deleteMyAgentToken
      description: |
        エージェント用トークンを失効させる（ApiDesign.md 4.5.5）。必要権限は「本人」。

        **行は消さず `revoked_at` を立てる**（4.4.3 と同じ）。

        **冪等。** 既に失効済みでも 204 を返し、`revoked_at` を上書きしない。

        **他人のエージェント・存在しない `id` / `token_id`・`token_type` が `agent`
        でないものは 404**（Design.md 6.4.5）。
      responses:
        '204':
          description: 失効させた（既に失効済みでも同じ）。本文を持たない。
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/CSRFFailed'
        '404':
          description: |
            自分のエージェントのトークンとして見つからない（`not_found`）。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pb_session
      description: |
        ブラウザ（Web UI）用。HttpOnly / SameSite=Lax（ApiDesign.md 2.3）。
        **状態変更系（POST/PATCH/PUT/DELETE）では `X-PB-CSRF` も必要**（2.4）。
    bearerAuth:
      type: http
      scheme: bearer
      description: 'CLI・スクリプト用（`Authorization: Bearer pb_api_...`。ApiDesign.md 2.3）。'
...
  responses:
...
    Unauthenticated:
      description: 未認証・トークン失効・アカウント無効（`unauthenticated`）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: レート制限（`rate_limited`。ApiDesign.md 2.9）。
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        X-RateLimit-Limit:
          $ref: '#/components/headers/XRateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/XRateLimitRemaining'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: サーバ内部エラー（`internal_error`）。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    AccessTokenList:
      description: |
        ApiDesign.md 4.4.1。ページネーションも ETag も持たない（1人5本が上限で、
        絞り込みも差分取得も意味を持たないため。7.1 / 7.2 と同じ扱い）。
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AccessToken'
...
    IssuedAccessToken:
      description: |
        発行直後の応答（ApiDesign.md 4.4.2）。**`token` を持つ唯一の形**で、
        AccessToken に平文を1項目足したものである。
      type: object
      required: [id, name, token, token_prefix, scopes, issued_at, status]
      properties:
        id:
          type: string
        name:
          type: string
          examples: ['CLI (MacBook)']
        token:
          description: |
            **平文。この応答でしか得られない。** 画面は1回だけ全文を出す
            （GuiDesign.md 5.8.1）。
          type: string
          examples: ['pb_api_9f3c1d2e...']
        token_prefix:
          type: string
          examples: ['pb_api_9']
        scopes:
          type: array
          items:
            type: string
        issued_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [active, expired]
      additionalProperties: false
    CreateTokenRequest:
      description: ApiDesign.md 4.4.2。
      type: object
      required: [name, expires_in_days]
      properties:
        name:
          description: |
            1〜100文字（`project.name` と同じ上限。`access_token.name` に DB の
            CHECK は無く、アプリ側が持つ）。前後の空白は取り除かれる。
          type: string
          minLength: 1
          maxLength: 100
          examples: ['CLI (MacBook)']
        expires_in_days:
          description: |
            1〜365。**無期限は許さない**——失効操作でしか消えないトークンが残り、
            置き忘れを検出できなくなるため（Design.md 6.5 がエージェントトークンに
            課す「有効期限必須」と揃える）。
          type: integer
          minimum: 1
          maximum: 365
          examples: [90]
        scopes:
          description: |
            権限カタログのキー（Design.md 6.4.2 の30件）。カタログに無い値は 422。
            省略時と空配列は「絞り込みなし」。**Phase 1 の画面は常に空で送る。**
          type: array
          items:
            type: string
          examples: [[]]
      additionalProperties: false
...
security:
  - cookieAuth: []
    csrfToken: []
  - bearerAuth: []